Code correction

+++ 0x15-file_io/1-create_file.c
@@ -13,8 +13,8 @@
 int create_file(const char *filename, char *text_content)
 {
 	int file_descriptor;
-	ssize_t text_len;
-	ssize_t bytes_written;
+	int text_len;
+	int bytes_written;
 
 	if (filename == NULL)
 		return (-1);
